Nice new program storing mechanism. TODO: Reading registers, there are 16 now.

+++ main.c
@@ -101,7 +101,7 @@ int main(int argc, char* argv[]) {
     }
     // Generate machine code
 
-    uint16_t* code = generate_code(instructions, instruction_count);
+    uint8_t * code = generate_code(instructions, instruction_count);
     //uint16_t* code_len = get_code_len(instructions);
 
     // Write the machine code to the output file
